+++ sregistry/main/google_drive/__init__.py
@@ -35,7 +35,6 @@ from oauth2client import client as oclient
 from .utils import ( create_folder, get_or_create_folder )
 from .pull import pull
 from .push import push
-from .record import record
 from .share import share
 from .query import ( container_query, list_containers, search, search_all )
 
@@ -121,7 +120,6 @@ class Client(ApiConnection):
 # Add your different functions imported at the top to the client here
 Client.pull = pull
 Client.push = push
-Client.record = record
 Client.share = share
 
 Client._get_or_create_folder = get_or_create_folder