+++ sregistry/main/google_storage/push.py
@@ -61,17 +61,10 @@ def push(self, path, name, tag=None):
     metadata = metadata['data']
     metadata.update(names)
 
-    #TODO: use multiple threads/other to upload
     manifest = self._upload(source=path, 
                             destination=names['storage'],
                             metadata=metadata)
 
-    # If result is successful, save container record
-    if manifest is not None:
-        metadata.update(manifest)
-        container = self.add(image_uri=names['uri'],
-                             metadata=metadata,
-                             url=manifest['mediaLink'])
     print(manifest['mediaLink'])